feat: improving required perms
