+++ discord/interactions/deleteConfigCommand.ts
@@ -8,9 +8,8 @@ import {
 } from "discord.js";
 
 import { DiscordServerConfigRepository } from "../../db";
+import { assertManageRoles } from "../../utils/discord/permissions";
 import { getRoleName } from "../role";
-
-import { assertManageRoles } from "./permissions";
 
 export const handleDeleteConfigCommand = async (
   interaction: ChatInputCommandInteraction,
